+++ src/Simpletools/Db/Cassandra/Schema.php
@@ -15,6 +15,8 @@ class Schema
 
 		self::$_schema[$cluster][$keyspace][$table]['schema'] = array();
 		self::$_schema[$cluster][$keyspace][$table]['primaryKey'] = array();
+		self::$_schema[$cluster][$keyspace][$table]['clusteringKey'] = array();
+		self::$_schema[$cluster][$keyspace][$table]['partitionKey'] = array();
 
 		$schema   = $client->connector()->schema();
 		$keyspaceObj = $schema->keyspace($keyspace);
@@ -40,9 +42,20 @@ class Schema
 			self::$_schema[$cluster][$keyspace][$table]['schema'][$column->name()] = (string)$column->type();
 		}
 
+
+		foreach ($tableObj->partitionKey() as $column)
+		{
+			self::$_schema[$cluster][$keyspace][$table]['partitionKey'][] = $column->name();
+		}
+
 		foreach ($tableObj->primaryKey() as $column)
 		{
 			self::$_schema[$cluster][$keyspace][$table]['primaryKey'][] = $column->name();
+		}
+
+		foreach ($tableObj->clusteringKey() as $column)
+		{
+			self::$_schema[$cluster][$keyspace][$table]['clusteringKey'][] = $column->name();
 		}
 
 		return self::$_schema[$cluster][$keyspace][$table]['schema'];
@@ -60,6 +73,29 @@ class Schema
 		return self::$_schema[$cluster][$keyspace][$table]['primaryKey'];
 	}
 
+	public static function getPartitionKey(Client $client, $keyspace, $table)
+	{
+		$cluster = $client->getCluster();
+
+		if(isset(self::$_schema[$cluster][$keyspace][$table]['partitionKey']))
+			return self::$_schema[$cluster][$keyspace][$table]['partitionKey'];
+
+		self::getSchema($client, $keyspace, $table);
+
+		return self::$_schema[$cluster][$keyspace][$table]['partitionKey'];
+	}
+
+	public static function getClusteringKey(Client $client, $keyspace, $table)
+	{
+		$cluster = $client->getCluster();
+
+		if(isset(self::$_schema[$cluster][$keyspace][$table]['clusteringKey']))
+			return self::$_schema[$cluster][$keyspace][$table]['clusteringKey'];
+
+		self::getSchema($client, $keyspace, $table);
+
+		return self::$_schema[$cluster][$keyspace][$table]['clusteringKey'];
+	}
 
 
 }